Pico-blink-sdk support for RP2350 (#58)
Modify CMakeLists.txt to automatically adjust for target architecture 
changes. Guide the swiftc compiler to use the appropriate target based 
on the PICO_PLATFORM CMake variable, which defaults to PICO_BOARD if 
not explicitly set.

Add a target check for RISC-V to support compilation on RP2350 hardware 
in RISC-V mode. Compilation currently fails due to pico-sdk errors, 
similar to issues encountered on RP2040 with the new SDK versions.